+++ repository-test/src/main/java/tech/ydb/yoj/repository/test/RepositoryTest.java
@@ -3361,6 +3361,26 @@ public abstract class RepositoryTest extends RepositoryTestSupport {
     }
 
     @Test
+    public void bulkInserts() {
+        var id1 = new Bubble.Id("a", "b");
+        var id2 = new Bubble.Id("c", "d");
+
+        db.tx(() -> {
+            db.bubbles().bulkUpsert(
+                    List.of(new Bubble(id1, "oldA", "oldB", "oldC"), new Bubble(id2, "oldA", "oldB", "oldC")),
+                    BulkParams.DEFAULT
+            );
+        });
+
+        db.readOnly().run(() -> {
+            var first = this.db.bubbles().find(id1);
+            assertThat(first).isNotNull();
+            assertThat(first.getFieldA()).isEqualTo("oldA");
+            assertThat(this.db.bubbles().find(id2)).isNotNull();
+        });
+    }
+
+    @Test
     public void loggingMdcContextEvenOnException() {
         Logger log = LoggerFactory.getLogger("RepositoryTest");
